+++ test/ProjectOrigin.Registry.IntegrationTests/RabbitMqHttpClientTests.cs
@@ -2,9 +2,9 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 using FluentAssertions;
+using ProjectOrigin.Registry.IntegrationTests.Fixtures;
 using ProjectOrigin.Registry.MessageBroker;
 using ProjectOrigin.Registry.Options;
-using ProjectOrigin.TestUtils;
 using RabbitMQ.Client;
 using Xunit;
 using MsOptions = Microsoft.Extensions.Options.Options;
